- thanks, John for the hint with the unsigned short int!  Reduced   mem requirements for the adjacent_vertices matrices of all higher-order   elements through converting to short int - reduced # of virtual fn's in Elem by one: the higher_order_equiv_elem()   or so got moved into a static member function of Elem.  This also enables   to choose whether one wants to convert quad4->quad8 or quad4->quad9   through a simple bool option in MeshBase::all_second_order(bool) - reduced repeated code in adjacent_vertices through moving these short   int matrices to Hex, Quad, Prism etc - fixed GMV support for higher-order Prisms (enuff debugging for today... ;-)

+++ libmesh/include/cell_inf_hex.h
@@ -1,4 +1,4 @@
-// $Id: cell_inf_hex.h,v 1.3 2003-08-07 19:25:30 ddreyer Exp $
+// $Id: cell_inf_hex.h,v 1.4 2003-08-18 14:44:51 ddreyer Exp $
 
 // The Next Great Finite Element Library.
 // Copyright (C) 2002  Benjamin S. Kirk, John W. Peterson
@@ -141,7 +141,7 @@ protected:
    * this.  Matrix that tells which vertices define the location
    * of mid-side (or second-order) nodes.  
    */
-  static const unsigned int _second_order_adjacent_vertices[8][2];
+  static const unsigned short int _second_order_adjacent_vertices[8][2];
   
 
 private:
