+++ libmesh/include/cell_inf_hex16.h
@@ -1,4 +1,4 @@
-// $Id: cell_inf_hex16.h,v 1.15 2003-08-07 19:25:30 ddreyer Exp $
+// $Id: cell_inf_hex16.h,v 1.16 2003-08-18 14:44:51 ddreyer Exp $
 
 // The Next Great Finite Element Library.
 // Copyright (C) 2002  Benjamin S. Kirk, John W. Peterson
@@ -125,8 +125,8 @@ public:
    * that defines the \f$ n^{th} \f$ second-order node.
    * Note that \p n is counted as depicted above, \f$ 8 \le n < 16 \f$.
    */
-  unsigned int second_order_adjacent_vertex (const unsigned int n,
-					     const unsigned int v) const;
+  unsigned short int second_order_adjacent_vertex (const unsigned int n,
+						   const unsigned int v) const;
